Move pie chart class, automatically set label color
This add a utility class to set the label color based on the brightness
of the background color, so that we don't need to add a bunch more
specific values to the switch statement.
